PK integrals Yoshimine sorting works with symmetry.
Like in the previous version of the code, only the pq and rs pairs
both in the totally symmetric representation are considered. This means
that if you have a density matrix that is not totally symmetric, this is
not the right code for you (yet).
Also, asymmetric density matrices still only work with C1 symmetry and
duct tape for now.
This is still a debug version of the code. Multithreading to add, and
testing to be done.